Refactor authentication code to use configuration settings and update dependencies
